NXP-26261: enhance security on comments

+++ nuxeo-features/nuxeo-platform-comment/nuxeo-platform-comment/src/main/java/org/nuxeo/ecm/platform/comment/impl/AnnotationServiceImpl.java
@@ -25,6 +25,8 @@ import static org.nuxeo.ecm.platform.comment.api.AnnotationConstants.ANNOTATION_
 import static org.nuxeo.ecm.platform.comment.api.AnnotationConstants.ANNOTATION_XPATH_PROPERTY;
 import static org.nuxeo.ecm.platform.comment.api.CommentManager.Feature.COMMENTS_LINKED_WITH_PROPERTY;
 import static org.nuxeo.ecm.platform.comment.api.ExternalEntityConstants.EXTERNAL_ENTITY_FACET;
+import static org.nuxeo.ecm.platform.comment.workflow.utils.CommentsConstants.COMMENT_AUTHOR;
+import static org.nuxeo.ecm.platform.comment.workflow.utils.CommentsConstants.COMMENT_PARENT_ID;
 import static org.nuxeo.ecm.platform.query.nxql.CoreQueryAndFetchPageProvider.CORE_SESSION_PROPERTY;
 
 import java.io.Serializable;
@@ -35,16 +37,21 @@ import java.util.stream.Collectors;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.nuxeo.ecm.core.api.CoreInstance;
 import org.nuxeo.ecm.core.api.CoreSession;
 import org.nuxeo.ecm.core.api.DocumentModel;
+import org.nuxeo.ecm.core.api.DocumentNotFoundException;
 import org.nuxeo.ecm.core.api.DocumentRef;
 import org.nuxeo.ecm.core.api.IdRef;
+import org.nuxeo.ecm.core.api.NuxeoPrincipal;
+import org.nuxeo.ecm.core.api.security.SecurityConstants;
 import org.nuxeo.ecm.platform.comment.api.Annotation;
 import org.nuxeo.ecm.platform.comment.api.AnnotationService;
 import org.nuxeo.ecm.platform.comment.api.CommentManager;
 import org.nuxeo.ecm.platform.comment.api.Comments;
 import org.nuxeo.ecm.platform.comment.api.ExternalEntity;
 import org.nuxeo.ecm.platform.comment.api.exceptions.CommentNotFoundException;
+import org.nuxeo.ecm.platform.comment.api.exceptions.CommentSecurityException;
 import org.nuxeo.ecm.platform.query.api.PageProvider;
 import org.nuxeo.ecm.platform.query.api.PageProviderService;
 import org.nuxeo.runtime.api.Framework;
@@ -62,87 +69,145 @@ public class AnnotationServiceImpl extends DefaultComponent implements Annotatio
     protected static final String GET_ANNOTATIONS_FOR_DOC_PAGEPROVIDER_NAME = "GET_ANNOTATIONS_FOR_DOCUMENT";
 
     @Override
-    public Annotation createAnnotation(CoreSession session, Annotation annotation) {
-        // Create base comment in the annotation
-        DocumentModel docToAnnotate = session.getDocument(new IdRef(annotation.getParentId()));
-        DocumentModel annotationModel = session.createDocumentModel(ANNOTATION_DOC_TYPE);
-        Comments.annotationToDocumentModel(annotation, annotationModel);
-        if (annotation instanceof ExternalEntity) {
-            annotationModel.addFacet(EXTERNAL_ENTITY_FACET);
-            Comments.externalEntityToDocumentModel((ExternalEntity) annotation, annotationModel);
+    public Annotation createAnnotation(CoreSession session, Annotation annotation) throws CommentSecurityException {
+        String parentId = annotation.getParentId();
+        if (!session.hasPermission(new IdRef(parentId), SecurityConstants.READ)) {
+            throw new CommentSecurityException("The user " + session.getPrincipal().getName()
+                    + " can not create annotations on document " + parentId);
         }
-        annotationModel = Framework.getService(CommentManager.class).createComment(docToAnnotate, annotationModel);
-        return Comments.newAnnotation(annotationModel);
+        return CoreInstance.doPrivileged(session, s -> {
+            // Create base comment in the annotation
+            DocumentModel docToAnnotate = s.getDocument(new IdRef(annotation.getParentId()));
+            DocumentModel annotationModel = s.createDocumentModel(ANNOTATION_DOC_TYPE);
+            Comments.annotationToDocumentModel(annotation, annotationModel);
+            if (annotation instanceof ExternalEntity) {
+                annotationModel.addFacet(EXTERNAL_ENTITY_FACET);
+                Comments.externalEntityToDocumentModel((ExternalEntity) annotation, annotationModel);
+            }
+            annotationModel = Framework.getService(CommentManager.class).createComment(docToAnnotate, annotationModel);
+            return Comments.newAnnotation(annotationModel);
+        });
     }
 
     @Override
-    public Annotation getAnnotation(CoreSession session, String annotationId) throws CommentNotFoundException {
+    public Annotation getAnnotation(CoreSession session, String annotationId)
+            throws CommentNotFoundException, CommentSecurityException {
         DocumentRef annotationRef = new IdRef(annotationId);
         if (!session.exists(annotationRef)) {
             throw new CommentNotFoundException("The document " + annotationId + " does not exist.");
         }
-        DocumentModel annotationModel = session.getDocument(annotationRef);
-        return Comments.newAnnotation(annotationModel);
+        NuxeoPrincipal principal = session.getPrincipal();
+        return CoreInstance.doPrivileged(session, s -> {
+            DocumentModel annotationModel = s.getDocument(annotationRef);
+            String parentId = (String) annotationModel.getPropertyValue(COMMENT_PARENT_ID);
+            if (!s.hasPermission(principal, new IdRef(parentId), SecurityConstants.READ)) {
+                throw new CommentSecurityException("The user " + principal.getName()
+                        + " does not have access to the annotations of document " + parentId);
+            }
+            return Comments.newAnnotation(annotationModel);
+        });
     }
 
     @Override
     @SuppressWarnings("unchecked")
     public List<Annotation> getAnnotations(CoreSession session, String documentId, String xpath)
-            throws CommentNotFoundException {
+            throws CommentNotFoundException, CommentSecurityException {
         DocumentRef docRef = new IdRef(documentId);
         if (!session.exists(docRef)) {
             throw new CommentNotFoundException("The document " + documentId + " does not exist.");
+        }
+        if (!session.hasPermission(docRef, SecurityConstants.READ)) {
+            throw new CommentSecurityException("The user " + session.getPrincipal().getName()
+                    + " does not have access to the annotations of document " + documentId);
         }
         DocumentModel annotatedDoc = session.getDocument(docRef);
         CommentManager commentManager = Framework.getService(CommentManager.class);
         if (commentManager.hasFeature(COMMENTS_LINKED_WITH_PROPERTY)) {
             PageProviderService ppService = Framework.getService(PageProviderService.class);
-            Map<String, Serializable> props = Collections.singletonMap(CORE_SESSION_PROPERTY, (Serializable) session);
-            PageProvider<DocumentModel> pageProvider = (PageProvider<DocumentModel>) ppService.getPageProvider(
-                    GET_ANNOTATIONS_FOR_DOC_PAGEPROVIDER_NAME, null, null, null, props, documentId, xpath);
-            return pageProvider.getCurrentPage().stream().map(Comments::newAnnotation).collect(Collectors.toList());
+            return CoreInstance.doPrivileged(session, s -> {
+                Map<String, Serializable> props = Collections.singletonMap(CORE_SESSION_PROPERTY, (Serializable) s);
+                PageProvider<DocumentModel> pageProvider = (PageProvider<DocumentModel>) ppService.getPageProvider(
+                        GET_ANNOTATIONS_FOR_DOC_PAGEPROVIDER_NAME, null, null, null, props, documentId, xpath);
+                return pageProvider.getCurrentPage().stream().map(Comments::newAnnotation).collect(Collectors.toList());
+            });
         }
-        return Framework.getService(CommentManager.class)
-                        .getComments(session, annotatedDoc)
-                        .stream()
-                        .filter(annotationModel -> ANNOTATION_DOC_TYPE.equals(annotationModel.getType())
-                                && xpath.equals(annotationModel.getPropertyValue(ANNOTATION_XPATH_PROPERTY)))
-                        .map(Comments::newAnnotation)
-                        .collect(Collectors.toList());
+        return CoreInstance.doPrivileged(session, s -> {
+            return commentManager
+                    .getComments(s, annotatedDoc)
+                    .stream()
+                    .filter(annotationModel -> ANNOTATION_DOC_TYPE.equals(annotationModel.getType())
+                            && xpath.equals(annotationModel.getPropertyValue(ANNOTATION_XPATH_PROPERTY)))
+                    .map(Comments::newAnnotation)
+                    .collect(Collectors.toList());
+        });
     }
 
     @Override
     public void updateAnnotation(CoreSession session, String annotationId, Annotation annotation)
-            throws CommentNotFoundException {
+            throws CommentNotFoundException, CommentSecurityException  {
         IdRef annotationRef = new IdRef(annotationId);
         if (!session.exists(annotationRef)) {
             throw new CommentNotFoundException("The annotation " + annotationId + " does not exist.");
         }
-        DocumentModel annotationModel = session.getDocument(annotationRef);
-        Comments.annotationToDocumentModel(annotation, annotationModel);
-        if (annotation instanceof ExternalEntity) {
-            Comments.externalEntityToDocumentModel((ExternalEntity) annotation, annotationModel);
+        NuxeoPrincipal principal = session.getPrincipal();
+        if (!principal.isAdministrator() && !annotation.getAuthor().equals(principal.getName())) {
+            throw new CommentSecurityException("The user " + principal.getName()
+                    + " can not edit annotations of document " + annotation.getParentId());
         }
-        session.saveDocument(annotationModel);
+        CoreInstance.doPrivileged(session, s -> {
+            DocumentModel annotationModel = s.getDocument(annotationRef);
+            Comments.annotationToDocumentModel(annotation, annotationModel);
+            if (annotation instanceof ExternalEntity) {
+                Comments.externalEntityToDocumentModel((ExternalEntity) annotation, annotationModel);
+            }
+            s.saveDocument(annotationModel);
+        });
     }
 
     @Override
     public void deleteAnnotation(CoreSession session, String annotationId) throws CommentNotFoundException {
-        Framework.getService(CommentManager.class).deleteComment(session, annotationId);
+        NuxeoPrincipal principal = session.getPrincipal();
+        CoreInstance.doPrivileged(session, s -> {
+            try {
+                DocumentModel annotation = s.getDocument(new IdRef(annotationId));
+                String parentId = (String) annotation.getPropertyValue(COMMENT_PARENT_ID);
+                DocumentRef parentRef = new IdRef(parentId);
+                if (!principal.isAdministrator()
+                        && !annotation.getPropertyValue(COMMENT_AUTHOR).equals(principal.getName())
+                        && !s.hasPermission(principal, parentRef, SecurityConstants.EVERYTHING)) {
+                    throw new CommentSecurityException(
+                            "The user " + principal.getName() + " can not delete comments of document " + parentId);
+                }
+            } catch (DocumentNotFoundException e) {
+                throw new CommentNotFoundException(e);
+            }
+            Framework.getService(CommentManager.class).deleteComment(s, annotationId);
+        });
     }
 
     @Override
-    public Annotation getExternalAnnotation(CoreSession session, String entityId) throws CommentNotFoundException {
+    public Annotation getExternalAnnotation(CoreSession session, String entityId)
+            throws CommentNotFoundException, CommentSecurityException {
         DocumentModel annotationModel = getAnnotationModel(session, entityId);
         if (annotationModel == null) {
             throw new CommentNotFoundException("The external annotation " + entityId + " does not exist.");
+        }
+        String parentId = (String) annotationModel.getPropertyValue(COMMENT_PARENT_ID);
+        if (!session.hasPermission(new IdRef(parentId), SecurityConstants.READ)) {
+            throw new CommentSecurityException("The user " + session.getPrincipal().getName()
+                    + " does not have access to the annotations of document " + parentId);
         }
         return Comments.newAnnotation(annotationModel);
     }
 
     @Override
     public void updateExternalAnnotation(CoreSession session, String entityId, Annotation annotation)
-            throws CommentNotFoundException {
+            throws CommentNotFoundException, CommentSecurityException {
+        NuxeoPrincipal principal = session.getPrincipal();
+        if (!principal.isAdministrator() && !annotation.getAuthor().equals(principal.getName())) {
+            throw new CommentSecurityException("The user " + session.getPrincipal().getName()
+                    + " can not edit annotations of document " + annotation.getParentId());
+        }
         DocumentModel annotationModel = getAnnotationModel(session, entityId);
         if (annotationModel == null) {
             throw new CommentNotFoundException("The external annotation " + entityId + " does not exist.");
@@ -155,10 +220,19 @@ public class AnnotationServiceImpl extends DefaultComponent implements Annotatio
     }
 
     @Override
-    public void deleteExternalAnnotation(CoreSession session, String entityId) throws CommentNotFoundException {
+    public void deleteExternalAnnotation(CoreSession session, String entityId)
+            throws CommentNotFoundException, CommentSecurityException {
         DocumentModel annotationModel = getAnnotationModel(session, entityId);
         if (annotationModel == null) {
             throw new CommentNotFoundException("The external annotation " + entityId + " does not exist.");
+        }
+        NuxeoPrincipal principal = session.getPrincipal();
+        String parentId = (String) annotationModel.getPropertyValue(COMMENT_PARENT_ID);
+        if (!principal.isAdministrator()
+                && !annotationModel.getPropertyValue(COMMENT_AUTHOR).equals(principal.getName())
+                && !session.hasPermission(new IdRef(parentId), SecurityConstants.EVERYTHING)) {
+            throw new CommentSecurityException(
+                    "The user " + principal.getName() + " can not delete comments of document " + parentId);
         }
         Framework.getService(CommentManager.class).deleteComment(session, annotationModel.getId());
     }
